Crate lea_compiler

This module contains the compiler. The compiler parses source code and generates byte code suitable for execution by the VM.

The complation process can be divided into the following phases:

Parsing

Checking

Resolving

Linting / Optimizing

Emitting

This module provides several helper methods that will either wrap up the whole compilation process or stop after a step, to enable more advanced uses of the compiler.

Below is a brief description of what each phase does.

Parsing

This phase runs the generated Lea parser. It is generated at compile-time from the Parsing Expression Grammar (PEG) in lea.rustpeg with help of the rust-peg crate.

Checking

In check.rs, a Checker visitor is implemented, which runs over an AST and verifies that all statements and expressions are used in a valid context. Since the PEG parser has no state, it cannot do this (and it is better - though slower - to do it this way, since it separates syntax error and context-dependent errors).

For example, the checker will deny a break statement used outside of a loop and the varargs expression ... outside of a varargs function.

Resolving

The resolver implemented in resolve.rs transforms any named variable (VNamed in the AST) to a local, upvalue, or global reference. This makes it easy to see which identifier refers to which variable (might be useful for code highlighting in IDEs) and makes the emitter a bit simpler, since it doesn't have to figure out what an identifier means if it encounters it.

Linting / Optimizing

In this step, all configured Lints and Optimizers are run on the resolved AST. Internally, we make no distinction between a Lint and an Optimizer: Both take a Function node, do something to it, and return a transformed Function along with a list of warnings to be issued.

Most Lints will just return the unmodified function and generate some warnings, while most optimizers will modify the AST, but return no warnings. Of course, there might be exceptions.

Lints and Optimizers are implemented in the files inside the "transform" directory:

fold.rs implements a constant folding optimization that looks for operators that are applied to literals, applies them at compile-time, and replaces the node with the result. This is a basic optimization performed by most compilers (Lua does this too).

globalwrite.rs implements a Lint that will warn for all assignments to globals. Global variables are generally meant to hold the standard library functions, and not much more. User code should always prefer locals, and modules are meant to return the module when loaded, which should be assigned to a local by the code that required the module. If you still need to access a global (for example, when setting up a sandbox), you can do so explicitly by indexing _ENV.

deprecated_ops.rs is a Lint that emits a warning whenever a Lua operator is used that has an equivalent replacement in Lea.

Emitting

The byte code emitter takes a resolved AST and emits byte code for each function that can be executed by the VM.

Modules

transform
Contains AST transformers such as Linters and Optimizers that can be run on an AST and report warnings or mutate it.

Structs

CompileConfig
Defines options that can be used to tweak the compilation process (optimizations, lints, ...)

CompileOutput
Output of a successful compilation.

FnData
Function representation output by the compiler. Later converted to a FunctionProto. This owns all child functions (as FnData) and can be serialized.

Warning

Enums

CompileError
Kinds of errors that can occur when compiling code

Functions

apply_transforms
Applies all transforms in the given CompileConfig to a main function. Returns all warnings generated by the transforms. main must be resolved beforehand.

compile_expr
Compiles an expression into an FnData object that will return the result(s) of the expression to the caller.

compile_str
Compiles source code into an FnData instance. This executes the whole compilation process and emits byte code.

parse_and_check
Parses and checks the given source code (stops after compilation step 2). Returns a valid Function node on success.

parse_and_resolve
Parses, checks and resolves the given source code (stops after compilation step 3). Returns the resolved main function on success.

Type Definitions

CompileResult