Enum lea_compiler::CompileError

pub enum CompileError {
    ErrParse(ParseError),
    ErrCheck(Vec<CheckError>),
    ErrLint(Vec<Warning>),
    ErrEmit(Vec<EmitError>),
}

Kinds of errors that can occur when compiling code

Variants
ErrParse(ParseError)

The input source code could not be parsed

ErrCheck(Vec<CheckError>)

The checker encountered one or more problems

ErrLint(Vec<Warning>)

One or more Lints have issued warnings and were configured to cause a compilation error. Contains all warnings of all Lints that were configured to error.

ErrEmit(Vec<EmitError>)

Byte code emission failed. This can happen if an implementation limit is reached.

Methods

impl CompileError

fn format<W: Write>(&self, code: &str, source_name: &str, t: &mut Terminal<Output=W>) -> Result<()>