Function lea_compiler::compile_str

pub fn compile_str<'a>(code: &'a str, source_name: &str, conf: &CompileConfig) -> CompileResult<CompileOutput<'a>>

Compiles source code into an FnData instance. This executes the whole compilation process and emits byte code.

Note that this will not return the Function node on error, even if parsing succeeded.

Parameters

code - The source code to compile

source_name - The name of this source code. Used when reporting errors. This should be a file name or something similar.

conf - The CompileConfig to use for this compilation