Module lea_compiler::transform

Contains AST transformers such as Linters and Optimizers that can be run on an AST and report warnings or mutate it.

Modules

deprecated_ops
Warns when a deprecated Lua operator is used (such as ~= or not).

fold
Performs constant folding on literals

globalwrite
Linter that emits a warning for all writes to globals. Note that this doesn't warn for writes to anything else, so math.s = ... is fine, while math = ... would cause a warning.

Structs

TRANSFORMS

TRANSFORMS_COMPAT

TRANSFORMS_DEFAULT

Enums

LintMode
Specifies how a Lint's (or Optimizer's) emitted warnings should be handled by the compiler.

Type Definitions

Transform
Transform function used by Linters and Optimizers